pyproject.toml
[tool.poetry]
name = "novel-editor"
version = "0.9.0"
description = ""
authors = ["Yamahara Yoshihiro <yoshihiro@yamahara.email>"]
readme = "README.md"
packages = [{include = "novel_editor"}]

[tool.poetry.dependencies]
python = "<3.13,>=3.8"
wikipedia-api = "^0.6.0"
janome = "^0.5.0"
jaconv = "^0.3.4"
pyttsx4 = "^3.0.15"
pillow = "^10.3.0"
customtkinter = "^5.2.2"
numpy = [
  {version = ">=1.26", python = ">=3.9"},
  {version = "<1.25", python = "=3.8"}
]
pycirclize = "^1.4.0"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
